requirements.txt
asgiref==3.8.1
attrs==25.1.0
Django==5.1.6
django-filter==25.1
django-js-asset==3.0.1
django-mptt==0.15.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.8
factory_boy==3.3.3
Faker==36.1.1
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.7
packaging==24.2
pluggy==1.5.0
pytest==7.2.2
pytest-django==4.5.2
pytest-factoryboy==2.7.0
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
